refactor(cli): move the publish group into its own module
Third step of splitting `cli/main.py`, following the pattern the `sort` pilot
established in #914. `publish` moves to `geoparquet_io/cli/commands/publish.py`
with its two commands, `stac` and `upload`.

The five STAC helpers travel with the group: `_stac_impl`,
`_handle_stac_item`, `_handle_stac_collection`, `_check_output_stac_item` and
`_check_output_stac_collection` are used by `publish stac` alone, so per the
`cli/commands/__init__.py` rule they belong in the group's module rather than
in `_shared.py`. They stay in `cli/`; none of them moves to `core/`.

The moved lines are byte-identical apart from the group's decorator line
(`@cli.group()` -> `@click.group()`, which would otherwise need `cli.main` and
make the dependency cyclic). `main.py` imports the group and attaches it with
`cli.add_command(publish)` in the existing registration block, so
`from geoparquet_io.cli.main import publish` still resolves; its now-unused
`core.upload` imports are dropped.

Also updated: `tests/test_upload.py` patched `check_credentials` on `cli.main`,
where `publish upload` used to resolve it. The twelve `patch.object` calls now
name `cli.commands.publish`. (`test_write_memory_forwarding` and
`test_cli_api_call_parity_scaffold`, the two files #914 parameterised by owning
module, list no `publish` case, so neither needed a change.)

`tests/data/cli_surface.json` is unchanged, and `gpio publish --help`,
`gpio publish stac --help` and `gpio publish upload --help` render identically
before and after.
